Clean up D2 draw call interception and fix some glitches.

+++ src/d2dx/RenderContext.cpp
@@ -983,5 +983,5 @@ void RenderContext::UnclipCursor()
 
 float RenderContext::GetFrameTime() const
 {
-	return _frameTimeMs / 1000.0;
+	return (float)(_frameTimeMs / 1000.0);
 }
